// src/proto/keymap/v1/keymap.ts
export interface Keymap {
  /** Keymap name should follow "users/{user}/keymaps/{keymap}" pattern, e.g. users/1234/keymaps/1 */
  name: string;
  /**
   * (-- api-linter: core::0122::embedded-resource=disabled
   *     aip.dev/not-precedent: Action is a value object, not a separately managed resource --)
   * The actions.
   * One action may correspond to multiple keybindings. Then there would be multiple Action
   * with the same keybinding but different id.
   * Clients should merge Action that share the same `id` if necessary.
   */
  actions: Action[];
}

/** Action represents a single mapping from multiple key bindings to a logical action. */
export interface Action {
  /** Action name should follow "keymap/{keymap}/action/{name}" pattern, e.g. keymap/1/action/actions.editor.copy */
  name: string;
  /**
   * The key bindings that trigger the action.
   * e.g. ctrl+s, ctrl+shift+s
   */
  bindings: KeybindingReadable[];
  /** The action config */
  actionConfig:
    | ActionConfig
    | undefined;
  /** Optional, user-provided comment for this keymap. */
  comment: string;
}
// ...
export interface Keybinding {
  /** The sequence of key chords. */
  chords: KeyChord[];
}

/** KeybindingReadable represents a single keybinding with its readable string */
export interface KeybindingReadable {
  /** The key binding that triggers an action. */
  keyChords:
    | Keybinding
    | undefined;
  /** The readable key binding that triggers an action. e.g. "ctrl+s" */
  keyChordsReadable: string;
}

/** ActionConfig represents the configuration of an action. */
export interface ActionConfig {
  /** name of the action */
  displayName: string;
  /** description of the action */
  description: string;
  /**
   * category of the action
   * e.g. "Editor", "Terminal"
   * Has subcategories if category has ".", e.g. "Editor.Clipboard"
   */
  category: string;
  /** editor support information for this action */
  editorSupport: EditorSupport[];
  /** children action ids */
  children: string[];
}

/** EditorSupport represents whether an action is supported by a specific editor. */
export interface EditorSupport {
  /** The editor type */
  editorType: EditorType;
  /** Whether this action is supported by the editor */
  supported: boolean;
  /** Special note of why not supported or supportness */
  note: string;
}

function createBaseKeymap(): Keymap {
  return { name: "", actions: [] };
}

export const Keymap: MessageFns<Keymap> = {
  encode(message: Keymap, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== "") {
      writer.uint32(10).string(message.name);
    }
    for (const v of message.actions) {
      Action.encode(v!, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Keymap {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseKeymap();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.actions.push(Action.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Keymap {
    return {
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      actions: globalThis.Array.isArray(object?.actions) ? object.actions.map((e: any) => Action.fromJSON(e)) : [],
    };
  },

  toJSON(message: Keymap): unknown {
    const obj: any = {};
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.actions?.length) {
      obj.actions = message.actions.map((e) => Action.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Keymap>, I>>(base?: I): Keymap {
    return Keymap.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Keymap>, I>>(object: I): Keymap {
    const message = createBaseKeymap();
    message.name = object.name ?? "";
    message.actions = object.actions?.map((e) => Action.fromPartial(e)) || [];
    return message;
  },
};

function createBaseAction(): Action {
  return { name: "", bindings: [], actionConfig: undefined, comment: "" };
}

export const Action: MessageFns<Action> = {
  encode(message: Action, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== "") {
      writer.uint32(10).string(message.name);
    }
    for (const v of message.bindings) {
      KeybindingReadable.encode(v!, writer.uint32(18).fork()).join();
    }
    if (message.actionConfig !== undefined) {
      ActionConfig.encode(message.actionConfig, writer.uint32(26).fork()).join();
    }
    if (message.comment !== "") {
      writer.uint32(34).string(message.comment);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Action {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAction();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.bindings.push(KeybindingReadable.decode(reader, reader.uint32()));
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.actionConfig = ActionConfig.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.comment = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Action {
    return {
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      bindings: globalThis.Array.isArray(object?.bindings)
        ? object.bindings.map((e: any) => KeybindingReadable.fromJSON(e))
        : [],
      actionConfig: isSet(object.actionConfig)
        ? ActionConfig.fromJSON(object.actionConfig)
        : isSet(object.action_config)
        ? ActionConfig.fromJSON(object.action_config)
        : undefined,
      comment: isSet(object.comment) ? globalThis.String(object.comment) : "",
    };
  },

  toJSON(message: Action): unknown {
    const obj: any = {};
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.bindings?.length) {
      obj.bindings = message.bindings.map((e) => KeybindingReadable.toJSON(e));
    }
    if (message.actionConfig !== undefined) {
      obj.actionConfig = ActionConfig.toJSON(message.actionConfig);
    }
    if (message.comment !== "") {
      obj.comment = message.comment;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Action>, I>>(base?: I): Action {
    return Action.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Action>, I>>(object: I): Action {
    const message = createBaseAction();
    message.name = object.name ?? "";
    message.bindings = object.bindings?.map((e) => KeybindingReadable.fromPartial(e)) || [];
    message.actionConfig = (object.actionConfig !== undefined && object.actionConfig !== null)
      ? ActionConfig.fromPartial(object.actionConfig)
      : undefined;
    message.comment = object.comment ?? "";
    return message;
  },
};
// ...
export const KeybindingReadable: MessageFns<KeybindingReadable> = {
  encode(message: KeybindingReadable, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.keyChords !== undefined) {
      Keybinding.encode(message.keyChords, writer.uint32(10).fork()).join();
    }
    if (message.keyChordsReadable !== "") {
      writer.uint32(18).string(message.keyChordsReadable);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): KeybindingReadable {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseKeybindingReadable();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.keyChords = Keybinding.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.keyChordsReadable = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): KeybindingReadable {
    return {
      keyChords: isSet(object.keyChords)
        ? Keybinding.fromJSON(object.keyChords)
        : isSet(object.key_chords)
        ? Keybinding.fromJSON(object.key_chords)
        : undefined,
      keyChordsReadable: isSet(object.keyChordsReadable)
        ? globalThis.String(object.keyChordsReadable)
        : isSet(object.key_chords_readable)
        ? globalThis.String(object.key_chords_readable)
        : "",
    };
  },

  toJSON(message: KeybindingReadable): unknown {
    const obj: any = {};
    if (message.keyChords !== undefined) {
      obj.keyChords = Keybinding.toJSON(message.keyChords);
    }
    if (message.keyChordsReadable !== "") {
      obj.keyChordsReadable = message.keyChordsReadable;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<KeybindingReadable>, I>>(base?: I): KeybindingReadable {
    return KeybindingReadable.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<KeybindingReadable>, I>>(object: I): KeybindingReadable {
    const message = createBaseKeybindingReadable();
    message.keyChords = (object.keyChords !== undefined && object.keyChords !== null)
      ? Keybinding.fromPartial(object.keyChords)
      : undefined;
    message.keyChordsReadable = object.keyChordsReadable ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}